+++ lending/src/codatlending/models/operations/get_categorized_profit_and_loss_statement.py
@@ -10,7 +10,7 @@ from typing import Optional
 
 
 @dataclasses.dataclass
-class GetEnhancedBalanceSheetAccountsRequest:
+class GetCategorizedProfitAndLossStatementRequest:
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
     report_date: str = dataclasses.field(metadata={'query_param': { 'field_name': 'reportDate', 'style': 'form', 'explode': True }})
     r"""The date in which the report is created up to. Users must specify a specific date, however the response will be provided for the full month."""
@@ -22,7 +22,7 @@ class GetEnhancedBalanceSheetAccountsRequest:
 
 
 @dataclasses.dataclass
-class GetEnhancedBalanceSheetAccountsResponse:
+class GetCategorizedProfitAndLossStatementResponse:
     content_type: str = dataclasses.field()
     status_code: int = dataclasses.field()
     enhanced_financial_report: Optional[shared_enhancedfinancialreport.EnhancedFinancialReport] = dataclasses.field(default=None)