+++ lending/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [python v0.1.2] lending
 ### Releases
 - [PyPI v0.1.2] https://pypi.org/project/codat-lending/0.1.2 - lending
+
+## 2023-09-06 00:16:07
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.78.8 (2.96.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.1.3] lending
+### Releases
+- [PyPI v0.1.3] https://pypi.org/project/codat-lending/0.1.3 - lending